[ayc122] <refactor> reorganised smps code
