+++ SMPS_code/Bidirectional_Original.py
